Fix 64-bit pointer truncation in Mesa paths

+++ mesa/src/mesa/main/imports.h
@@ -99,7 +99,7 @@ extern "C" {
  * this macro.
  * Both pointers/offsets are expressed in bytes.
  */
-#define ADD_POINTERS(A, B)  ( (A) + (unsigned long) (B) )
+#define ADD_POINTERS(A, B)  ( (A) + (size_t) (B) )
 
 
 /**********************************************************************/
